Wokring on the Arudino
Created the first branch called arduino where i will begin to modify how
the arduino will behave as a data source in this project.

I have implemented a simple script for the Arduino that outlines
the basic setup and loop structure. The Arduino script (.src/main.cpp)
is designed to control two LEDs and a servo based on commands received via
serial communication. It initializes the LEDs and servo on specified pins
and listens for incoming serial commands to execute predefined actions.
Each command is expected to trigger a corresponding action, which
needs to be defined in the handleCommand function.

I have also implemented a structure to the readme file so i can easy
update everystep in furture implementation to the code so everything
can be documented and easy to follow for some one els intressted.