Created a function called handle_response
This is responsible for handling the response received from a server.
The function takes in one argument, response, which is expected to be a byte object representing
the response received from the server.
The purpose of this function is to validate and extract information from the response header.
The response header is assumed to be 3 bytes long and contains information about the response
type and the length of the data. The code first extracts the first 3 bytes of the response, which
is the header, using slicing. Then, it uses the struct.unpack function to unpack the header
bytes into two variables: response_type_validation and data_length.
The '>BH' format string specifies that the first byte should be interpreted as an unsigned char (B),
and the next two bytes should be interpreted as an unsigned short (H), with the '>' indicating big-endian byte order.

Next, the code attempts to create a MessageType object using the response_type_validation value.
If the value is not a valid response type, a ValueError is raised. If the MessageType object is successfully created,
the function continues with further processing.

If a ValueError is raised during the creation of the MessageType object, the function simply returns,
indicating that the response type is invalid.