wokring on a function to get a user command
This code snippet defines a function called get_user_command() that displays a menu of options for the user and prompts for their choice.
The function starts by printing a list of options for the user to choose from.
Each option is represented by a number and a corresponding action.
After printing the menu, the function prompts the user to enter their choice by using the input() function.
The user's choice is then returned as a string.